# Optimize Mobile renderer by using FP16 explicitly.