--- a/servers/rendering/renderer_rd/shaders/decal_data_inc.glsl
+++ b/servers/rendering/renderer_rd/shaders/decal_data_inc.glsl
@@ -1,17 +1,17 @@
 struct DecalData {
-	highp mat4 xform; //to decal transform
-	highp vec3 inv_extents;
-	mediump float albedo_mix;
-	highp vec4 albedo_rect;
-	highp vec4 normal_rect;
-	highp vec4 orm_rect;
-	highp vec4 emission_rect;
-	highp vec4 modulate;
-	mediump float emission_energy;
+	mat4 xform; //to decal transform
+	vec3 inv_extents;
+	float albedo_mix;
+	vec4 albedo_rect;
+	vec4 normal_rect;
+	vec4 orm_rect;
+	vec4 emission_rect;
+	vec4 modulate;
+	float emission_energy;
 	uint mask;
-	mediump float upper_fade;
-	mediump float lower_fade;
-	mediump mat3x4 normal_xform;
-	mediump vec3 normal;
-	mediump float normal_fade;
+	float upper_fade;
+	float lower_fade;
+	mat3x4 normal_xform;
+	vec3 normal;
+	float normal_fade;
 };