--- a/servers/rendering/rendering_shader_library.h
+++ b/servers/rendering/rendering_shader_library.h
@@ -36,6 +36,8 @@ public:
 		FEATURE_ADVANCED_BIT = 1U << 0U,
 		FEATURE_MULTIVIEW_BIT = 1U << 1U,
 		FEATURE_VRS_BIT = 1U << 2U,
+		FEATURE_FP16_BIT = 1U << 3U,
+		FEATURE_FP32_BIT = 1U << 4U,
 	};
 
 	// Used by the shader baker to globally enable features on all the shaders that will be exported.