… servers/rendering/renderer_rd/shaders/forward_mobile/scene_forward_mobile.glsl
… servers/rendering/renderer_rd/shaders/forward_mobile/scene_forward_mobile_inc.glsl
… servers/rendering/renderer_rd/shaders/half_inc.glsl
… servers/rendering/renderer_rd/shaders/light_data_inc.glsl
… servers/rendering/renderer_rd/shaders/scene_data_inc.glsl
… servers/rendering/renderer_rd/shaders/scene_forward_aa_inc.glsl
… servers/rendering/renderer_rd/shaders/scene_forward_lights_inc.glsl
… servers/rendering/renderer_rd/shaders/scene_forward_vertex_lights_inc.glsl
… servers/rendering/rendering_device_commons.h