Supersedes https://github.com/godotengine/godot/pull/99767.

This PR took a while to find the right way to achieve the results I was looking for, which is ultimately to achieve a frame time reduction, but FP16 is known to carry other types of improvements that might be harder to measure when it comes to power consumption.

The overall strategy has been:
- Introduce a new way of explicitly marking FP16 usage by introducing new types such as `half`, `hvec3` and others. This brings GLSL more in line with the methodology both HLSL and MSL apply when it comes to these types. While the changes might look pretty invasive at first glance, the explicit usage makes it much easier to tell what is going on and achieve actual results. 
  - This is not the first time FP16 usage has been attempted, as seen in https://github.com/godotengine/godot/pull/99767 (and https://github.com/godotengine/godot/pull/52500). However, this was a pretty clear demonstration that the level of control provided by mechanisms like `mediump` can be pretty difficult to work with, and that PR did not provide any actual performance benefits unlike this one.
  - Another key benefit is that SPIR-V Cross does not support relaxed precision when translating to other languages and discards the information. That is not the case with explicit FP16, which means Metal finally gets FP16 optimizations!
- The results from user shader code are explicitly converted to FP16 for variables that don't exactly warrant having such high precision. This led to large improvements in reducing register pressure. User shader code is maintained on full precision as the concept for FP16 is not exposed to user shader code at the moment.
- Most lighting code has been adapted and converted to use FP16 while avoiding errors in areas where high precision is necessary. Some parts of the code have not been adapted because converting between FP32 and FP16 was just not worth the cost.
- [ARM's Mali Offline Compiler](https://developer.arm.com/Tools%20and%20Software/Arm%20Performance%20Studio) was crucial to informing the impacts of each of these changes and whether it resulted in an improvement or not, as rapid prototyping was required and seeing the actual impact of the shader code change upon the final shader generation.
- FP16 shaders are selected for mobile based on the features reported by RenderingDevice. The goal is that users should not perceive any quality loss from using it, as it's been inserted in places where the precision loss shouldn't be significant. However, we should fix any issues that arise from doing this.

## Results

### Register pressure

Register usage has seen pretty massive reductions across the board, from around 30% to 50%, with instances of shaders that used several features going from suffering from spilling registers to the stack to not requiring it at all. Less usage of registers has also led to several instances of occupancy boosts.

### Measurable FP16 Utilization

On an M3 we can actually see a pretty sizable bump in FP16 utilization, while it was 0% before due to the way SPIR-V Cross works.

![image](https://github.com/user-attachments/assets/28619b70-441f-4985-84d7-cec969bf1f29)

### Average GPU Time

[TPS Demo from here](https://github.com/godotengine/godot/pull/90284)

**Mali-G78**
`master`: ~24.6 ms
`explicit-fp16`: ~24.2 ms

**Mali-G715**
`master`: ~17.3 ms
`explicit-fp16`: ~16.6 ms

The overall improvement will probably heavily vary depending on project content and how much can the hardware take advantage of the FP16 usage.

## Changes to Shader Baker

Adding this feature sadly requires bumping up the size of the shader container to include both FP16 and FP32 variants so the right one can be loaded based on hardware support. We might explore the possibility of making these optional toggles in case it turns out to be a problem.

### TODO

- [x] Thoroughly review that all user shader code has the right highp versions of variables available.
- [x] Solve build errors in Forward+.
- [x] Test for regressions in visual quality.

---

Contributed by [W4 Games](https://w4games.com/). 🍀