@DarioSamo 2025-06-04
Heads-up, there's currently a major visual regression in Forward+ which I'm currently tracking down, which seems to be related to one of the changes to the common lighting file.