@clayjohn 2025-06-04
I tested with the mobile renderer on an Adreno 640 device and cannot see any visual regressions. But I did note a consistent 10% improvement to the frame time in the TPS demo!